+++ src/Symfony/Component/Cache/Traits/MemcachedTrait.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Cache\Traits;
 
 use Symfony\Component\Cache\Exception\CacheException;
 use Symfony\Component\Cache\Exception\InvalidArgumentException;
+use Symfony\Component\Dsn\Factory\MemcachedFactory;
 
 /**
  * @author Rob Frawley 2nd <rmf@src.run>
@@ -22,13 +23,6 @@ use Symfony\Component\Cache\Exception\InvalidArgumentException;
  */
 trait MemcachedTrait
 {
-    private static $defaultClientOptions = array(
-        'persistent_id' => null,
-        'username' => null,
-        'password' => null,
-        'serializer' => 'php',
-    );
-
     private $client;
     private $lazyClient;
 
@@ -58,134 +52,38 @@ trait MemcachedTrait
     }
 
     /**
-     * Creates a Memcached instance.
-     *
-     * By default, the binary protocol, no block, and libketama compatible options are enabled.
-     *
-     * Examples for servers:
-     * - 'memcached://user:pass@localhost?weight=33'
-     * - array(array('localhost', 11211, 33))
-     *
-     * @param array[]|string|string[] An array of servers, a DSN, or an array of DSNs
-     * @param array                   An array of options
-     *
-     * @return \Memcached
-     *
-     * @throws \ErrorEception When invalid options or servers are provided
+     * @see MemcachedFactory::create()
      */
     public static function createConnection($servers, array $options = array())
     {
-        if (is_string($servers)) {
-            $servers = array($servers);
-        } elseif (!is_array($servers)) {
-            throw new InvalidArgumentException(sprintf('MemcachedAdapter::createClient() expects array or string as first argument, %s given.', gettype($servers)));
+        @trigger_error(sprintf('The %s() method is deprecated since version 3.4 and will be removed in 4.0. Use the MemcachedFactory::create() method from Dsn component instead.', __METHOD__), E_USER_DEPRECATED);
+
+        foreach ((array) $servers as $i => $dsn) {
+            if (is_array($dsn)) {
+                @trigger_error(sprintf('Passing an array of array to the %s() method is deprecated since version 3.4 and will be removed in 4.0. Use the MemcachedFactory::create() method from Dsn component instead with an array of Dsn instead.', __METHOD__), E_USER_DEPRECATED);
+
+                $scheme = 'memcached://';
+                $host = isset($dsn['host']) ? $dsn['host'] : '';
+                $port = isset($dsn['port']) ? ':'.$dsn['port'] : '';
+                $user = isset($dsn['user']) ? $dsn['user'] : '';
+                $pass = isset($dsn['pass']) ? ':'.$dsn['pass'] : '';
+                $pass = ($user || $pass) ? "$pass@" : '';
+                $path = isset($dsn['path']) ? $dsn['path'] : '';
+                $query = isset($dsn['query']) ? '?'.$dsn['query'] : '';
+                $fragment = isset($dsn['fragment']) ? '#'.$dsn['fragment'] : '';
+
+                $servers[$i] = "$scheme$user$pass$host$port$path$query$fragment";
+            }
         }
+
         if (!static::isSupported()) {
             throw new CacheException('Memcached >= 2.2.0 is required');
         }
-        set_error_handler(function ($type, $msg, $file, $line) { throw new \ErrorException($msg, 0, $type, $file, $line); });
+
         try {
-            $options += static::$defaultClientOptions;
-            $client = new \Memcached($options['persistent_id']);
-            $username = $options['username'];
-            $password = $options['password'];
-
-            // parse any DSN in $servers
-            foreach ($servers as $i => $dsn) {
-                if (is_array($dsn)) {
-                    continue;
-                }
-                if (0 !== strpos($dsn, 'memcached://')) {
-                    throw new InvalidArgumentException(sprintf('Invalid Memcached DSN: %s does not start with "memcached://"', $dsn));
-                }
-                $params = preg_replace_callback('#^memcached://(?:([^@]*+)@)?#', function ($m) use (&$username, &$password) {
-                    if (!empty($m[1])) {
-                        list($username, $password) = explode(':', $m[1], 2) + array(1 => null);
-                    }
-
-                    return 'file://';
-                }, $dsn);
-                if (false === $params = parse_url($params)) {
-                    throw new InvalidArgumentException(sprintf('Invalid Memcached DSN: %s', $dsn));
-                }
-                if (!isset($params['host']) && !isset($params['path'])) {
-                    throw new InvalidArgumentException(sprintf('Invalid Memcached DSN: %s', $dsn));
-                }
-                if (isset($params['path']) && preg_match('#/(\d+)$#', $params['path'], $m)) {
-                    $params['weight'] = $m[1];
-                    $params['path'] = substr($params['path'], 0, -strlen($m[0]));
-                }
-                $params += array(
-                    'host' => isset($params['host']) ? $params['host'] : $params['path'],
-                    'port' => isset($params['host']) ? 11211 : null,
-                    'weight' => 0,
-                );
-                if (isset($params['query'])) {
-                    parse_str($params['query'], $query);
-                    $params += $query;
-                    $options = $query + $options;
-                }
-
-                $servers[$i] = array($params['host'], $params['port'], $params['weight']);
-            }
-
-            // set client's options
-            unset($options['persistent_id'], $options['username'], $options['password'], $options['weight']);
-            $options = array_change_key_case($options, CASE_UPPER);
-            $client->setOption(\Memcached::OPT_BINARY_PROTOCOL, true);
-            $client->setOption(\Memcached::OPT_NO_BLOCK, true);
-            if (!array_key_exists('LIBKETAMA_COMPATIBLE', $options) && !array_key_exists(\Memcached::OPT_LIBKETAMA_COMPATIBLE, $options)) {
-                $client->setOption(\Memcached::OPT_LIBKETAMA_COMPATIBLE, true);
-            }
-            foreach ($options as $name => $value) {
-                if (is_int($name)) {
-                    continue;
-                }
-                if ('HASH' === $name || 'SERIALIZER' === $name || 'DISTRIBUTION' === $name) {
-                    $value = constant('Memcached::'.$name.'_'.strtoupper($value));
-                }
-                $opt = constant('Memcached::OPT_'.$name);
-
-                unset($options[$name]);
-                $options[$opt] = $value;
-            }
-            $client->setOptions($options);
-
-            // set client's servers, taking care of persistent connections
-            if (!$client->isPristine()) {
-                $oldServers = array();
-                foreach ($client->getServerList() as $server) {
-                    $oldServers[] = array($server['host'], $server['port']);
-                }
-
-                $newServers = array();
-                foreach ($servers as $server) {
-                    if (1 < count($server)) {
-                        $server = array_values($server);
-                        unset($server[2]);
-                        $server[1] = (int) $server[1];
-                    }
-                    $newServers[] = $server;
-                }
-
-                if ($oldServers !== $newServers) {
-                    // before resetting, ensure $servers is valid
-                    $client->addServers($servers);
-                    $client->resetServerList();
-                }
-            }
-            $client->addServers($servers);
-
-            if (null !== $username || null !== $password) {
-                if (!method_exists($client, 'setSaslAuthData')) {
-                    trigger_error('Missing SASL support: the memcached extension must be compiled with --enable-memcached-sasl.');
-                }
-                $client->setSaslAuthData($username, $password);
-            }
-
-            return $client;
-        } finally {
-            restore_error_handler();
+            return MemcachedFactory::create($servers, $options);
+        } catch (\Symfony\Component\Dsn\Exception\InvalidArgumentException $e) {
+            throw new InvalidArgumentException($e->getMessage(), 0, $e);
         }
     }